# contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settings/notifications": {
      "get": {
        "operationId": "SettingsController_getAlertNotifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateAlertNotifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertNotificationsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/settings/sso": {
      "get": {
        "operationId": "SettingsController_getSsoConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateSsoConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSsoConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/settings/quota": {
      "get": {
        "operationId": "SettingsController_getQuotaUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/settings/branding": {
      "get": {
        "operationId": "SettingsController_getBranding",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateBranding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "settings"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EasySignage API",
    "description": "REST do CMS (JWT), rotas públicas de pareamento e API do player (Bearer de device). Especificação evolutiva — alinhar com `digital_signage_arquitetura_roadmap.md`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "device-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateAlertNotificationsDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSsoConfigDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBrandingDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}